// client/src/models/Property.enum.ts
export enum AdditionalInfo {
    BASEMENT = "BASEMENT",
    BALCONY = "BALCONY",
    ELEVATOR = "ELEVATOR",
    OWN_BATH = "OWN_BATH",
    STOVE = 'STOVE',
    OWN_KITCHEN = 'OWN_KITCHEN',
    OWN_TOILET = 'OWN_TOILET',
    REFRIGERATOR = 'REFRIGERATOR',
    PETS_ALLOWED = 'PETS_ALLOWED',
    SHARED_LAUNDRY = 'SHARED_LAUNDRY'
}